requirements.txt
streamlit
requests
websocket-client